Move N_PREFIX to the start of PATH
Sometimes packages require Node as a dependency so we may be forced to install it with Homebrew. But we want to use the version(s) installed with n. Putting N_PREFIX at the start of the PATH means we can prefer items installed via NPM over Homebrew.

I don't have this issue at the moment but I changed it in case I do in the future.